requirements.txt
flask
gunicorn
Pillow
requests
